Remove startup parameter from JOB command
The _ZJOB routine adds a startup parameter to the JOB command that
executes gtmenv to set basic environment variables and basic environment
sanity, however YottaDB/GT.M doesn't support setting environment
variables for the JOBed process on Linux, so the startup parameter
is essentially useless. The _ZJOB routine also doesn't provide a
full path to the startup parameter, so it is unable to execute the
parameter which causes the JOB command to fail.

Since it is unnecessary to set the environment variables again,
remove the setting of the startup parameter.