Initial version of FileMan to PIP mapping tools
This includes the MASH Unit Test Framework as a dependency (licensed
Apache 2.0).

Initial version of the FileMan to PIP tools supports most fields
that are part of the parent file. Some support for subfiles is
included, but not fully tested and doesn't have test cases written.

Unit tests are written for the PATIENT (File #2) and NEW PERSON
(File #200) Files/Tables and every test passes except for one where
the Field name is duplicated with a "*".