[#17] Fix recursive SubFile number retrieval
When nested SubFiles were involved the mapper wouldn't traverse the
parents correctly and would return an invalid key structure. This
was resolved by modifying the algorithm to look at the parent before
the SubFile we are trying to map to get the right SubFile number
to add the correct key.

Perform clean up of comments and other lines so they fit
better in an editor window.

Standardize messages that are printed to the screen with Error or
Info.